LayoutTests/http/tests/frames/frames-with-same-name-expected.txt

Make sure that frame lookup by name has correct ordering when duplicate names are involved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.foo is document.getElementById('firstFrame').contentWindow
PASS open('', 'foo') is window.foo
PASS successfullyParsed is true

TEST COMPLETE
